change files names to prober ones
