+++ packages/testsweets/lib/src/setup_code.dart
@@ -1,5 +1,5 @@
 import 'package:testsweets/src/locator.dart';
 
-Future<void> setupTestSweets() async{
-  await setupLocator(); 
+Future<void> setupTestSweets() async {
+  await setupLocator();
 }
