+++ packages/testsweets/lib/src/ui/widget_capture/widget_capture_view.dart
@@ -2,19 +2,18 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked/stacked_annotations.dart';
-import 'package:testsweets/src/constants/app_constants.dart';
 import 'package:testsweets/src/enums/capture_widget_enum.dart';
 import 'package:testsweets/src/ui/shared/app_colors.dart';
 import 'package:testsweets/src/ui/shared/cta_button.dart';
 import 'package:testsweets/src/ui/shared/shared_styles.dart';
 import 'package:testsweets/src/ui/widget_capture/widget_capture_view.form.dart';
 import 'package:testsweets/src/ui/widget_capture/widget_capture_viewmodel.dart';
-import 'package:testsweets/src/ui/widget_capture/widget_capture_widgets/widget_description_dialog_widget.dart';
+import 'package:testsweets/src/ui/widget_capture/widget_capture_widgets/widget_description_dialog.dart';
 
-import 'widget_capture_widgets/capture_view_layout.dart';
-import 'widget_capture_widgets/inspect_layout_widget.dart';
-import 'widget_capture_widgets/main_view_layout.dart';
-import 'widget_capture_widgets/widget_name_input.dart';
+import 'widget_capture_widgets/capture_controllers.dart';
+import 'widget_capture_widgets/capture_layout.dart';
+import 'widget_capture_widgets/inspect_controllers.dart';
+import 'widget_capture_widgets/intro_controllers.dart';
 
 @FormView(fields: [FormTextField(name: 'widgetName')])
 class WidgetCaptureView extends StatelessWidget with $WidgetCaptureView {
@@ -44,18 +43,12 @@ class WidgetCaptureView extends StatelessWidget with $WidgetCaptureView {
                       builder: (context) => Stack(
                             children: [
                               child,
-                              if (model.captureWidgetStatusEnum ==
-                                      CaptureWidgetStatusEnum.inspectMode ||
-                                  model.captureWidgetStatusEnum ==
-                                      CaptureWidgetStatusEnum
-                                          .inspectModeDialogShow)
-                                InspectLayoutView(),
-                              if (model.captureWidgetStatusEnum ==
-                                      CaptureWidgetStatusEnum.captureMode ||
-                                  model.captureWidgetStatusEnum ==
-                                      CaptureWidgetStatusEnum
-                                          .captureModeWidgetNameInputShow)
-                                _CaptureLayout(
+                              if (model.captureWidgetStatusEnum
+                                  .isAtInspectModeMode())
+                                InspectControllers(),
+                              if (model.captureWidgetStatusEnum
+                                  .isAtCaptureMode())
+                                CaptureLayout(
                                     widgetNameController: widgetNameController,
                                     widgetNameFocusNode: widgetNameFocusNode),
                               Positioned(
@@ -70,7 +63,7 @@ class WidgetCaptureView extends StatelessWidget with $WidgetCaptureView {
                                       children: [
                                         if (model.captureWidgetStatusEnum ==
                                             CaptureWidgetStatusEnum.idle)
-                                          MainViewLayout(),
+                                          IntroControllers(),
                                         if (model.captureWidgetStatusEnum ==
                                             CaptureWidgetStatusEnum.inspectMode)
                                           CtaButton(
@@ -87,7 +80,7 @@ class WidgetCaptureView extends StatelessWidget with $WidgetCaptureView {
                                             model.captureWidgetStatusEnum ==
                                                 CaptureWidgetStatusEnum
                                                     .captureModeAddWidget)
-                                          CaptureViewLayout(),
+                                          CaptureControllers(),
                                       ],
                                     ),
                                   )),
@@ -131,72 +124,6 @@ class WidgetCaptureView extends StatelessWidget with $WidgetCaptureView {
                 ],
               )),
       viewModelBuilder: () => WidgetCaptureViewModel(projectId: projectId),
-    );
-  }
-}
-
-class _CaptureLayout extends ViewModelWidget<WidgetCaptureViewModel> {
-  const _CaptureLayout({
-    Key? key,
-    required this.widgetNameController,
-    required this.widgetNameFocusNode,
-  }) : super(key: key);
-
-  final TextEditingController widgetNameController;
-  final FocusNode widgetNameFocusNode;
-
-  @override
-  Widget build(BuildContext context, WidgetCaptureViewModel model) {
-    return Stack(
-      children: [
-        if (
-        // model.hasWidgetDescription &&
-        model.captureWidgetStatusEnum ==
-            CaptureWidgetStatusEnum.captureModeWidgetNameInputShow) ...[
-          Positioned(
-              top: model.descriptionTop,
-              left: model.descriptionLeft,
-              child: GestureDetector(
-                onPanUpdate: (panEvent) {
-                  final x = panEvent.delta.dx;
-                  final y = panEvent.delta.dy;
-                  model.updateDescriptionPosition(x, y);
-                },
-                child: Container(
-                  width: WIDGET_DESCRIPTION_VISUAL_SIZE,
-                  height: WIDGET_DESCRIPTION_VISUAL_SIZE,
-                  decoration: BoxDecoration(
-                    color: Colors.pink,
-                    borderRadius: BorderRadius.circular(20),
-                  ),
-                ),
-              )),
-        ],
-        FadeInWidget(
-            child: AnimatedAlign(
-              duration: const Duration(milliseconds: 500),
-              curve: Curves.easeOutCubic,
-              alignment: model.widgetNameInputPositionIsDown
-                  ? Alignment.bottomCenter
-                  : Alignment.topCenter,
-              widthFactor: 1,
-              child: WidgetNameInput(
-                errorMessage: model.nameInputErrorMessage,
-                closeWidget: () {
-                  widgetNameController.clear();
-                  model.closeWidgetNameInput();
-                },
-                saveWidget: model.saveWidgetDescription,
-                switchPositionTap: model.switchWidgetNameInputPosition,
-                focusNode: widgetNameFocusNode,
-                textEditingController: widgetNameController,
-              ),
-            ),
-            isVisible: !model.isBusy &&
-                // model.hasWidgetDescription &&
-                model.captureWidgetStatusEnum ==
-                    CaptureWidgetStatusEnum.captureModeWidgetNameInputShow),
-      ],
     );
   }
 }
