+++ packages/testsweets/lib/src/ui/widget_capture/widget_capture_widgets/capture_controllers.dart
@@ -3,12 +3,12 @@ import 'package:stacked/stacked.dart';
 import 'package:testsweets/src/enums/capture_widget_enum.dart';
 import 'package:testsweets/src/ui/shared/app_colors.dart';
 import 'package:testsweets/src/ui/shared/cta_button.dart';
-import 'package:testsweets/src/ui/widget_capture/widget_capture_widgets/widgets_container.dart';
+import 'package:testsweets/src/ui/widget_capture/widget_capture_widgets/widget_types_container.dart';
 
 import '../widget_capture_viewmodel.dart';
 
-class CaptureViewLayout extends ViewModelWidget<WidgetCaptureViewModel> {
-  const CaptureViewLayout({
+class CaptureControllers extends ViewModelWidget<WidgetCaptureViewModel> {
+  const CaptureControllers({
     Key? key,
   }) : super(key: key);
 
@@ -29,7 +29,7 @@ class CaptureViewLayout extends ViewModelWidget<WidgetCaptureViewModel> {
             duration: const Duration(milliseconds: 350),
             child: model.captureWidgetStatusEnum ==
                     CaptureWidgetStatusEnum.captureModeWidgetsContainerShow
-                ? WidgetsContainer()
+                ? WidgetsTypesContainer()
                 : CtaButton(
                     title: 'Add Widget',
                     fillColor: kcPassedTestGreenColor,
