--- packages/testsweets/lib/src/ui/widget_capture/widget_capture_widgets/widget_description_capture_layer.dart
@@ -1,29 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:stacked/stacked.dart';
-
-import '../widget_capture_viewmodel.dart';
-
-/// The widget that sits at the top and allows you to tap on screen and add a [WidgetDescription]
-/// to the view
-class WidgetDescriptionCaptureLayer
-    extends ViewModelWidget<WidgetCaptureViewModel> {
-  const WidgetDescriptionCaptureLayer({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context, WidgetCaptureViewModel model) {
-    return Positioned.fill(
-        child: Container(
-      width: double.infinity,
-      height: double.infinity,
-      alignment: Alignment.bottomRight,
-      padding: const EdgeInsets.all(15),
-      decoration: BoxDecoration(
-          border: Border.all(
-        color: Colors.blue,
-        width: 5,
-      )),
-    ));
-  }
-}
